+++ ot_simple_rest/ot_simple_rest.py
@@ -14,12 +14,13 @@ from psycopg2.pool import ThreadedConnectionPool
 from handlers.eva.auth import AuthLoginHandler
 from handlers.eva.logs import LogsHandler
 from handlers.eva.dashs import DashboardHandler, DashboardsHandler, SvgLoadHandler, DashExportHandler, \
-    DashImportHandler, GroupExportHandler, GroupImportHandler
+    DashImportHandler, GroupExportHandler, GroupImportHandler, DashByNameHandler
 from handlers.eva.quizs import QuizsHandler, QuizHandler, QuizQuestionsHandler, QuizFilledHandler, \
     FilledQuizExportHandler, QuizExportJsonHandler, QuizImportJsonHandler, CatalogsListHandler, CatalogHandler
 from handlers.eva.role_model import UserHandler, UsersHandler, RoleHandler, RolesHandler, \
     PermissionsHandler, PermissionHandler, GroupsHandler, GroupHandler, UserPermissionsHandler, \
     IndexesHandler, IndexHandler, UserGroupsHandler, UserDashboardsHandler, GroupDashboardsHandler, UserSettingHandler
+from handlers.eva.papers import PaperLoadHandler, PapersHandler, PaperHandler
 
 from handlers.jobs.makejob import MakeJob
 from handlers.jobs.loadjob import LoadJob
@@ -37,7 +38,7 @@ __author__ = "Andrey Starchenkov"
 __copyright__ = "Copyright 2019, Open Technologies 98"
 __credits__ = ["Anton Khromov"]
 __license__ = ""
-__version__ = "1.5.2"
+__version__ = "1.6.0"
 __maintainer__ = "Andrey Starchenkov"
 __email__ = "astarchenkov@ot.ru"
 __status__ = "Production"
@@ -132,6 +133,7 @@ def main():
 
     logger = set_logger(config['general'].get('level', 'INFO'),
                         os.path.join(base_logs_dir, 'otsimplerest.log'), 'osr')
+    logger.info('Version: %s' % __version__)
     logger.info('DB configuration: %s' % db_conf)
     logger.info('MEM configuration: %s' % mem_conf)
 
@@ -191,6 +193,8 @@ def main():
         (r'/api/dash', DashboardHandler, {"db_conn_pool": db_pool_eva}),
         (r'/api/dash/export', DashExportHandler, {"db_conn_pool": db_pool_eva, "static_conf": static_conf}),
         (r'/api/dash/import', DashImportHandler, {"db_conn_pool": db_pool_eva}),
+        (r'/api/dashByName', DashByNameHandler, {"db_conn_pool": db_pool_eva}),
+        
 
         (r'/api/load/svg', SvgLoadHandler, {"db_conn_pool": db_pool_eva, "static_conf": static_conf}),
 
@@ -213,6 +217,10 @@ def main():
         (r'/qapi/catalog/create', CatalogHandler, {"db_conn_pool": db_pool_eva}),
         (r'/qapi/catalog/edit', CatalogHandler, {"db_conn_pool": db_pool_eva}),
         (r'/qapi/catalog/delete', CatalogHandler, {"db_conn_pool": db_pool_eva}),
+
+        (r'/api/eva/reports/load', PaperLoadHandler, {"db_conn_pool": db_pool_eva,"static_conf": static_conf}),
+        (r'/api/eva/reports/getAll', PapersHandler, {"db_conn_pool": db_pool_eva,"static_conf": static_conf}),
+        (r'/api/eva/reports/get', PaperHandler, {"db_conn_pool": db_pool_eva,"static_conf": static_conf,"mem_conf": mem_conf}),
     ],
         login_url=r'/api/auth/login'
     )