+++ tests/rest/eva_tester.py
@@ -446,6 +446,17 @@ class EvaTester:
             self._cleanup()
         return dash_from_api['data']['name'] == data['name']
 
+    def test__get_dash_by_name(self):
+        data = {'name': 'test_dash'}
+        try:
+            self.send_request(method='POST', endpoint='/api/dash', data=data)
+            dash_from_db = self.db.execute_query('SELECT id, name FROM dash WHERE name=%s;',
+                                                 params=(data['name'],), as_obj=True)
+            dash_from_api = self.send_request(method='GET', endpoint=f'/api/dashByName?name={dash_from_db.name}&idgroup=0')
+        finally:
+            self._cleanup()
+        return dash_from_api['data']['name'] == data['name']
+
     def test__get_dashs_list(self):
         data = [{'name': 'dash_1'}, {'name': 'dash_2'}]
         try: